(#85) handle STAN connection errors
In the past should the STAN connection be interrupted long enough
the whole system would just stop working as on reconnect the server
would not recognise us as valid.

In the go library version 0.4.0 and the server version 0.10.0 a
bidirectional ping was added whereby the client code can get notified on
disconnection.  The problem is one has to entirely recreate all
subscriptions and everything from scratch when this happens.  Quite a
difficult ask for a program that was never developed with that in mind.

Thus to shoe horn this pattern in here we set up a channel where each of
the STAN connections can notify the main manager code who then intiates
a clean shutdown via cancel() and a full fresh restart via Exec() which
would in effect close everything and reopen everything cleanly, it's a
bit brutal as stats will get lost but it gets us somewhere quickly till
we later consider rearchitecting all of this around the new programming
model STAN demands :(