(#85) handle reconnects
In the past should the STAN connection be interrupted long enough
the whole system would just stop working as on reconnect the server
would not recognise us as valid.

In the go library version 0.4.0 and the server version 0.10.0 a
bidirectional ping was added whereby the client code can get notified on
disconnection.  The problem is one has to entirely recreate all
subscriptions and everything from scratch when this happens.  Quite a
difficult ask for a program that was never developed with that in mind.

To use this we refactor the connector class so it's fully responsible
for interacting with STAN rather than the weird uncontained split we
had before and now in that class we handle reconnects and resubscribes